+++ src-tauri/src/manager/downloader.rs
@@ -292,7 +292,7 @@ impl<'a> Installer<'a> {
             }
         }
 
-        if installer::try_cache_install(data, &path, &mut manager, &thunderstore)? {
+        if installer::try_cache_install(data, &path, &mut manager, &thunderstore, &prefs)? {
             self.completed_bytes += borrowed.version.file_size;
             save(&manager, &prefs)?;
             return Ok(InstallMethod::Cached);
@@ -367,7 +367,7 @@ impl<'a> Installer<'a> {
             callback(install, &mut manager, &thunderstore);
         }
 
-        installer::try_cache_install(install, &path, &mut manager, &thunderstore)
+        installer::try_cache_install(install, &path, &mut manager, &thunderstore, &prefs)
             .context("failed to install after download")?;
 
         manager