+++ src/lib/components/Checkbox.svelte
@@ -17,7 +17,7 @@
 		class="rounded-md w-6 h-6 p-1 
           bg-{value ? 'green-700' : 'gray-800'}
           hover:bg-{value ? 'green-600' : 'gray-700'}
-          {value ? '' : 'border border-gray-500'}"
+          {!value && 'border border-gray-500'}"
 	>
 		{#if value}
 			<Icon class="text-white w-full h-full font-bold" icon="mdi:check" />